src/restaurant-service/requirements.txt
google-cloud-pubsub==2.2.0
Flask==2.0.2 # Required for Cloud Functions HTTP trigger context
